Added testing by using pytest

+++ routers/todos.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from starlette import status
 from models import Todos
 from db import SessionLocal
+from routers.auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -28,31 +30,40 @@ class TodoRequest(BaseModel):
 
 
 @router.get('/')
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Auth failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get("user_id")).all()
 
 
 @router.get('/todos/{todo_id}', status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Auth failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found')
 
 
 @router.post('/todo/', status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.dict())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Auth failed")
+    todo_model = Todos(**todo_request.dict(), owner_id=user.get("user_id"))
     db.add(todo_model)
     db.commit()
 
 
 @router.put('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
-                    todo_request: TodoRequest,
+async def update_todo(user: user_dependency,
+                      db: db_dependency,
+                      todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)
                       ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Auth failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
@@ -67,10 +78,12 @@ async def update_todo(db: db_dependency,
 
 
 @router.delete('/todo/delete/{todo_id}/', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Auth failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).delete()
 
     db.commit()
